Вынес вспомогательные функции добавления в корзину с хук. поправил документацию
